refactor: reference new apis
